# services/hub-api/openapi.yaml
openapi: 3.1.0
info:
  title: MK Hub API
  version: 0.1.0
  description: |
    Central API for the MK Platform — manages units, bookings, tickets,
    and Beds24 integration for both COBNB and Monthly Key brands.
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/v1/tickets:
    get:
      summary: List ops tickets
      tags: [Tickets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, IN_PROGRESS, COMPLETED, CANCELLED] }
        - name: type
          in: query
          schema: { type: string, enum: [CLEANING, MAINTENANCE, INSPECTION, TURNOVER] }
      responses:
        "200":
          description: Ticket list
    post:
      summary: Create ops ticket
      tags: [Tickets]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTicket" }
      responses:
        "201":
          description: Ticket created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTicket:
      type: object
      properties:
        unitId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        type: { type: string, enum: [CLEANING, MAINTENANCE, INSPECTION, TURNOVER] }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
        description: { type: string }
      required: [unitId, type, priority]